+++ welly/quality.py
@@ -1,8 +1,7 @@
-# -*- coding: utf-8 -*-
 """
 Quality functions for welly.
 
-:copyright: 2016 Agile Geoscience
+:copyright: 2021 Agile Scientific
 :license: Apache 2.0
 """
 import numpy as np